feat: add Vite plugin configuration for PWA and Sentry
- implement base Vite plugins for improved functionality
- add PWA and Sentry plugin support for production and preview modes
- refactor Vite config to utilize new plugin functions
