+++ config/plugins/index.ts
@@ -1,0 +1,4 @@
+export { getBaseVitePlugins, getBaseWorkerPlugins } from './base';
+export { getPwaPlugins } from './pwa';
+export { getSentryPlugins } from './sentry';
+export { getSslPlugins } from './ssl';
