+++ config/plugins/ssl.ts
@@ -1,0 +1,10 @@
+import type { PluginOption } from 'vite';
+import basicSsl from '@vitejs/plugin-basic-ssl';
+
+type GetSslPluginsParams = {
+  isPreview: boolean;
+  mode: string;
+};
+
+export const getSslPlugins = ({ isPreview, mode }: GetSslPluginsParams): PluginOption[] =>
+  mode === 'development' || isPreview ? [basicSsl()] : [];
